added data directory
Kaggle data is too large to put on github.  I have a link to where you can download it in a readme file, as well as the jupyter notebook that needs it.  here I just have a symbolic link to my dropbox folder that will only work on my local machine.  We should figure out a more global solution.